Fix windows build
Modify which files and how they are included. Based on the CMake
include directories copied the problematic template class.
This violates DRY and should be followed by refactoring.
Update appveyor Qt versions to 5.12.3.
