+++ src/ipc/facelift-ipc.h
@@ -31,7 +31,7 @@
 #pragma once
 
 #include "IPCServiceAdapterBase.h"
-#include "IPCProxyBase.h"
+#include "IPCProxyNewBase.h"
 
 namespace facelift {
