+++ src/ipc/ipc-common/CMakeLists.txt
@@ -2,12 +2,12 @@
 facelift_add_library(FaceliftIPCCommonLib
     SOURCES
         ipc-common.cpp
-        IPCProxyBase.cpp
+        IPCProxyNewBase.cpp
         ModuleIPCBase.cpp
         ipc-serialization.cpp
     HEADERS
         ipc-common.h
-        IPCProxyBase.h
+        IPCProxyNewBase.h
         IPCServiceAdapterBase.h
         ModuleIPCBase.h
         ipc-serialization.h
