+++ src/ipc/ipc-common/IPCProxyNewBase.h
