+++ src/ipc/ipc-common/ipc-common.h
@@ -281,99 +281,6 @@ inline void assignDefaultValue(Type &v)
     v = Type {};
 }
 
-template<typename AdapterType>
-class IPCProxyBase : public AdapterType
-{
-
-public:
-    using InterfaceType = AdapterType;
-
-public:
-    IPCProxyBase(QObject *parent) : AdapterType(parent)
-    {
-    }
-
-    template<typename BinderType>
-    void initBinder(BinderType &binder)
-    {
-        m_ipcBinder = &binder;
-        QObject::connect(this, &InterfaceBase::componentCompleted, &binder, &BinderType::onComponentCompleted);
-    }
-
-    bool isSynchronous() const
-    {
-        return m_ipcBinder->isSynchronous();
-    }
-
-    bool ready() const override final
-    {
-        auto r = m_serviceReady;
-        return r;
-    }
-
-    void setServiceReady(bool isServiceReady)
-    {
-        if (ready() != isServiceReady) {
-            m_serviceReady = isServiceReady;
-            emit this->readyChanged();
-        }
-    }
-
-    virtual void emitChangeSignals()
-    {
-        emit this->readyChanged();
-    }
-
-    void deserializeCommonSignal(facelift::CommonSignalID signalID)
-    {
-        switch (signalID) {
-        case facelift::CommonSignalID::readyChanged:
-            emit this->readyChanged();
-            break;
-        default:
-            qFatal("Unknown signal ID");
-        }
-    }
-
-    template<typename ProxyType>
-    class InterfacePropertyIPCProxyHandler
-    {
-
-    public:
-        InterfacePropertyIPCProxyHandler(IPCProxyBase &owner) : m_owner(owner)
-        {
-        }
-
-        void update(const QString &objectPath)
-        {
-            if (m_proxy && (m_proxy->ipc()->objectPath() != objectPath)) {
-                m_proxy = nullptr;
-            }
-            if (!m_proxy) {
-                m_proxy = m_owner.m_ipcBinder->template getOrCreateSubProxy<ProxyType>(objectPath);
-            }
-        }
-
-        ProxyType *getValue() const
-        {
-            return m_proxy;
-        }
-
-    private:
-        QPointer<ProxyType> m_proxy;
-        IPCProxyBase &m_owner;
-    };
-
-
-private:
-    bool m_serviceReady = false;
-
-protected:
-    IPCProxyBinderBase *m_ipcBinder = nullptr;
-
-};
-
-
 class FaceliftIPCCommonLib_EXPORT IPCAttachedPropertyFactoryBase : public QObject
 {
     Q_OBJECT
