+++ src/ipc/local/CMakeLists.txt
@@ -8,6 +8,7 @@ facelift_add_library(FaceliftIPCLocalLib
         LocalIPCRegistry.h
         LocalIPC-serialization.h
         LocalIPCServiceAdapter.h
+        IPCProxyBase.h
     LINK_LIBRARIES FaceliftIPCCommonLib
     PUBLIC_HEADER_BASE_PATH ${CMAKE_CURRENT_SOURCE_DIR}
 )
